requirements.txt
MaaFw==3.0.0
numpy==2.0.2
pyinstaller==6.9.0
pyside6==6.8.1.1
loguru==0.7.3